fix: move token-related functions to auth.js in utils folder
