+++ api/src/routers/logout.js
@@ -1,14 +1,14 @@
 import express from 'express'
-import { cleanupInvalidatedTokens } from '../utils/cleanupInvalidatedTokens.js'
-import { invalidateAuthToken } from '../utils/invalidateAuthToken.js'
+import { cleanupInvalidatedTokens } from '../utils/auth.js'
+import { invalidateAuthToken } from '../utils/auth.js'
 
 const logout = express.Router()
 
 logout.post('/', async (req, res) => {
-  const token = req.headers['authorization']?.split(' ')[1]
+  const token = req.header('Authorization')?.split(' ')[1]
 
   if (!token) {
-    return res.status(400).json({ message: 'No token provided' })
+    return res.status(401).json({ message: 'No token provided' })
   }
 
   try {
